Add content handlers and built-in SSE support
Content handlers are a chain of modules implementing callbacks
that receive the body of responses and may modify it (for example
for decompressing the content) or act upon it (like sending a
message to the owner process.

The gun_sse content handler module can be used to translate
text/event-stream events on the fly and deliver them to the
owner process as a {gun_sse...} message.

This feature is currently not documented and is only tested
against a public server. It requires an up to date Cowlib.