Remove explicit call to dbg:start/0
This fixes a bug in `gun` when trying to open 2 connections with tracing enabled.

You can recreate it like so:
    > application:ensure_all_started(gun).
    > gun:open("localhost", 80, #{trace => true}).
    > gun:open("localhost", 80, #{trace => true}).

The second call fails with a cause clause exception.

You'll observe the same exception by:
   > dbg:start().
   > dbg:start().

Ultimately this may be a bug in OTP, but since `dbg:start/0` is undocumented
I'm inclined to give it the benefit of the doubt. Regardless, since the docs[1]
clearly state that calling `dbg:tracer/0` is the proper way to start the default
tracer message receiver, I argue that calling `dbg:start/0` is a bug in `gun`.

[1]: http://erlang.org/doc/man/dbg.html